chore: MovesTo method with map attributes
- Renames "CanMove" to "MovesTo" in order to make more sense (the method wasn't a boolean).
- Adds additional values to MapAttribute (OutOfBounds, Player, Event) and sorts them all alphabetically.
- MovesTo now returns updated MapAttribute values from the core instead of confusing integers, which allows us to have more readable and simplified code.
- frmMapLayers.cs: removes Method 'GetAttributeFromEditor' which contained unnecessary int casts to map attributes plus, is never used.
- Updates Dash's CalculateRange with the new values and a simplified switch.

+++ Intersect.Server/Maps/MapInstance.cs
@@ -466,7 +466,7 @@ namespace Intersect.Server.Maps
                 {
                     x = (byte)Randomization.Next(0, Options.MapWidth);
                     y = (byte)Randomization.Next(0, Options.MapHeight);
-                    if (mMapController.Attributes[x, y] == null || mMapController.Attributes[x, y].Type == (int)MapAttribute.Walkable)
+                    if (mMapController.Attributes[x, y] == null || mMapController.Attributes[x, y].Type == MapAttribute.Walkable)
                     {
                         break;
                     }
